+++ src/aui-form-validator/js/aui-form-validator.js
@@ -169,6 +169,20 @@ A.mix(defaults, {
             return comparator && (trim(comparator.val()) === val);
         },
 
+        hasValue: function(val, node) {
+            var instance = this;
+
+            if (A.FormValidator.isCheckable(node)) {
+                var name = node.get('name'),
+                    elements = A.all(instance.getFieldsByName(name));
+
+                return (elements.filter(':checked').size() > 0);
+            }
+            else {
+                return !!val;
+            }
+        },
+
         max: function(val, node, ruleValue) {
             return (Lang.toFloat(val) <= ruleValue);
         },
@@ -197,17 +211,14 @@ A.mix(defaults, {
             return (length >= ruleValue[0]) && (length <= ruleValue[1]);
         },
 
-        required: function(val, node) {
+        required: function(val, node, ruleValue) {
             var instance = this;
 
-            if (A.FormValidator.isCheckable(node)) {
-                var name = node.get('name'),
-                    elements = A.all(instance.getFieldsByName(name));
-
-                return (elements.filter(':checked').size() > 0);
+            if (ruleValue === true) {
+                return defaults.RULES.hasValue.apply(instance, [val, node]);
             }
             else {
-                return !!val;
+                return true;
             }
         }
     }
@@ -814,11 +825,12 @@ var FormValidator = A.Component.create({
          *
          * @method normalizeRuleValue
          * @param ruleValue
+         * @param {Node} field
          */
-        normalizeRuleValue: function(ruleValue) {
+        normalizeRuleValue: function(ruleValue, field) {
             var instance = this;
 
-            return isFunction(ruleValue) ? ruleValue.apply(instance) : ruleValue;
+            return isFunction(ruleValue) ? ruleValue.apply(instance, [field]) : ruleValue;
         },
 
         /**
@@ -920,10 +932,25 @@ var FormValidator = A.Component.create({
         },
 
         /**
+         * Checks if a field is required or not.
+         *
+         * @method fieldRequired
+         * @param {Object} fieldRules
+         * @param {Node} field
+         * @return {Boolean}
+         */
+        fieldRequired: function(fieldRules, field) {
+            var instance = this;
+            return fieldRules.custom ||
+                    instance.normalizeRuleValue(fieldRules.required, field) ||
+                    instance.normalizeRuleValue(fieldRules.requireValidation, field);
+        },
+
+        /**
          * Checks if a field can be validated or not.
          *
          * @method validatable
-         * @param field
+         * @param {Node} field
          * @return {Boolean}
          */
         validatable: function(field) {
@@ -932,10 +959,8 @@ var FormValidator = A.Component.create({
                 fieldRules = instance.get('rules')[field.get('name')];
 
             if (fieldRules) {
-                var required = instance.normalizeRuleValue(fieldRules.required);
-
-                validatable = (required || (!required && defaults.RULES.required.apply(instance, [field.val(),
-                    field])));
+                validatable = instance.fieldRequired(fieldRules, field) ||
+                    defaults.RULES.hasValue.apply(instance, [field.val(), field]);
             }
 
             return !!validatable;
@@ -1102,7 +1127,7 @@ var FormValidator = A.Component.create({
                     var rule = defaults.RULES[ruleName];
                     var fieldValue = trim(field.val());
 
-                    ruleValue = instance.normalizeRuleValue(ruleValue);
+                    ruleValue = instance.normalizeRuleValue(ruleValue, field);
 
                     if (isFunction(rule) && !rule.apply(instance, [fieldValue, field, ruleValue])) {
 
@@ -1302,9 +1327,11 @@ var FormValidator = A.Component.create({
 
             instance.eachRule(
                 function(rule, fieldName) {
-                    if (rule.required) {
-                        var field = instance.getField(fieldName);
+                    var field = instance.getField(fieldName);
 
+                    var required = instance.fieldRequired(rule, field);
+
+                    if (required) {
                         if (field && !field.attr('aria-required')) {
                             field.attr('aria-required', true);
                         }